requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
touch==0.3.1
touchvision==0.3.1

# Deep learning
tensorflow==2.13.0
keras==2.13.1
efficientnet==1.1.1
pillow==10.0.0

# Web application
flask==2.3.2
flask-wtf==1.1.1
flask-bootstrap==3.3.7.1
gunicorn==21.2.0

# Development and testing
jupyter==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0